ENH: Adding function to create a new WidgetRepresentation for a module
This can allow the user to "import" a module's GUI into another.
This is particularly interesting for work-flow modules.
The function getModuleGui can grab the existing GUI for the given module,
but the widget representation can only exist in one place. So when using
the GUI in another module with getModuleGui would cause one of the modules
to be empty.